+++ src/client/java/rh/maparthelper/conversion/CroppingMode.java
@@ -1,7 +1,8 @@
 package rh.maparthelper.conversion;
 
 public enum CroppingMode {
-    NO_CROP,
     AUTO_CROP,
+    NO_CROP,
+    FIT,
     USER_CROP
 }