+++ src/client/java/rh/maparthelper/conversion/ImageChangeResult.java
@@ -1,0 +1,7 @@
+package rh.maparthelper.conversion;
+
+public enum ImageChangeResult {
+    SIMPLE,
+    NEED_RESCALE,
+    TOP_LINE_CHANGED
+}