+++ src/client/java/rh/maparthelper/conversion/NativeImageUtils.java
@@ -4,18 +4,22 @@ import net.minecraft.client.MinecraftClient;
 import net.minecraft.client.texture.NativeImage;
 import net.minecraft.client.texture.NativeImageBackedTexture;
 import net.minecraft.client.texture.TextureManager;
+import rh.maparthelper.colors.MapColorEntry;
+import rh.maparthelper.conversion.mapart.ConvertedMapartImage;
 
 import java.awt.image.BufferedImage;
+import java.awt.image.DataBufferInt;
 
 public class NativeImageUtils {
 
-    public static void updateMapartImageTexture(NativeImage image) {
-        CurrentConversionSettings.guiMapartImage = new NativeImageBackedTexture(
+    public synchronized static void updateMapartImageTexture(NativeImage image) {
+        NativeImageBackedTexture backedTexture = new NativeImageBackedTexture(
                 () -> "mapart_gui_texture",
                 image
         );
         TextureManager textureManager = MinecraftClient.getInstance().getTextureManager();
-        textureManager.registerTexture(CurrentConversionSettings.guiMapartId, CurrentConversionSettings.guiMapartImage);
+        textureManager.registerTexture(CurrentConversionSettings.guiMapartId, backedTexture);
+        CurrentConversionSettings.guiMapartImage = backedTexture;
     }
 
     public static int[][] divideMapartByMaps(ConvertedMapartImage mapart) {
@@ -25,7 +29,7 @@ public class NativeImageUtils {
         int height = mapart.getHeight();
         int imageWidth = width * 128;
 
-        int[] pixels = mapart.image.copyPixelsArgb();
+        int[] pixels = mapart.getNativeImage().copyPixelsArgb();
         int[][] maps = new int[width * height][];
         for (int i = 0; i < maps.length; i++) {
             maps[i] = new int[16384];
@@ -42,16 +46,32 @@ public class NativeImageUtils {
         return maps;
     }
 
-    public static NativeImage convertBufferedImageToNativeImage(BufferedImage image) {
+    public static NativeImage convertBufferedImageToNativeImage(BufferedImage image, MapColorEntry bgColor, boolean useTransparent) {
         int width = image.getWidth();
         int height = image.getHeight();
 
+        int[] pixels = ((DataBufferInt) image.getRaster().getDataBuffer()).getData();
         NativeImage nativeImage = new NativeImage(width, height, false);
 
         for (int y = 0; y < height; y++) {
             for (int x = 0; x < width; x++) {
-                int argb = image.getRGB(x, y);
-                nativeImage.setColorArgb(x, y, argb);
+                int argb = pixels[x + y * width];
+
+                if (useTransparent ? argb == 0 : ((argb >> 24) & 0xFF) < 80) {
+                    nativeImage.setColorArgb(x, y, bgColor.getRenderColor());
+                } else {
+                    nativeImage.setColorArgb(x, y, useTransparent ? argb : argb | 0xFF000000);
+                }
+
+                if (useTransparent) {
+                    if (argb == 0) continue;
+                    nativeImage.setColorArgb(x, y, argb);
+                } else {
+                    if (((argb >> 24) & 0xFF) < 80)
+                        nativeImage.setColorArgb(x, y, bgColor.getRenderColor());
+                    else
+                        nativeImage.setColorArgb(x, y, argb | 0xFF000000);
+                }
             }
         }