+++ src/main/java/rh/maparthelper/conversion/staircases/Waves3DStaircase.java
@@ -1,10 +1,6 @@
 package rh.maparthelper.conversion.staircases;
 
-import net.minecraft.block.MapColor;
-import rh.maparthelper.config.palette.PaletteColors;
-
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
 
@@ -12,11 +8,11 @@ public class Waves3DStaircase implements IMapartStaircase {
 
     @Override
     public List<List<Integer>> getStaircase(int[][] colors) {
+        int[][] brights = getBrightnesses(colors);
         int height = colors.length + 1;
         int width = colors[0].length;
 
         int[] negativeHeights = new int[width];
-        Arrays.fill(negativeHeights, 0);
 
         List<List<Integer>> staircase = new ArrayList<>();
         for (int z = 0; z < height; z++) {
@@ -27,7 +23,7 @@ public class Waves3DStaircase implements IMapartStaircase {
         for (int x = 0; x < width; x++) {
             for (int z = height - 2; z >= 0; z--) {
                 if (z > 0 && colors[z - 1][x] == 0) continue;
-                int y = staircase.get(z + 1).get(x) + getHeightAlt(colors, x, z);
+                int y = staircase.get(z + 1).get(x) + getHeightAlt(brights, x, z);
                 staircase.get(z).set(x, y);
                 negativeHeights[x] = Math.min(negativeHeights[x], y);
             }
@@ -45,11 +41,11 @@ public class Waves3DStaircase implements IMapartStaircase {
         return staircase;
     }
 
-    private static int getHeightAlt(int[][] colors, int x, int z) {
-        if (z == colors.length) return 0;
-        MapColor.Brightness brightness = PaletteColors.getMapColorEntryByARGB(colors[z][x]).brightness();
-        if (brightness == MapColor.Brightness.LOW) return 1;
-        if (brightness == MapColor.Brightness.HIGH) return -1;
+    private static int getHeightAlt(int[][] brights, int x, int z) {
+        if (z == brights.length) return 0;
+        int brightness = brights[z][x];
+        if (brightness == 0) return 1;
+        if (brightness == 2) return -1;
         return 0;
     }
 }